requirements.txt
numpy==1.21.6
pandas==1.3.5
xlrd==2.0.1
google-cloud-storage==2.4.0
google-api-python-client==2.54.0
